[#42] moved data related model and API to v1_data from v1_form
