+++ backend/api/v1/v1_data/apps.py
@@ -1,0 +1,6 @@
+from django.apps import AppConfig
+
+
+class V1DataConfig(AppConfig):
+    default_auto_field = 'django.db.models.BigAutoField'
+    name = 'api.v1.v1_data'
