+++ backend/api/v1/v1_data/constants.py
@@ -1,0 +1,10 @@
+class DataApprovalStatus:
+    pending = 1
+    approved = 2
+    rejected = 3
+
+    FieldStr = {
+        pending: 'Pending',
+        approved: 'Approved',
+        rejected: 'Rejected',
+    }
